Add `pooling.arrangement`
